Rename the GpuDrawItemHeader struct to GpuDrawItem
Before this commit, client code using the GpuDevice module accesses
pointers to GpuDrawItem, which is a forward-declared type
that doesn't have a full definition. Internally, the GpuDevice module
always casts a GpuDrawItem* to a GpuDrawItemHeader* (which is the actual
struct).

Here, we remove the need for this casting by simply renaming
GpuDrawItemHeader to GpuDrawItem.