Reduce size of GpuDrawItem struct
Make the GpuDrawItem struct smaller by using u16's instead of u32's
for resource IDs (i.e. IDs for pipeline state objects and buffers).

The resource IDs (e.g. a GpuBufferID), which are 32-bit, contain
both the actual index of the resource in the lookup table, but
also an 'inner ID'. This inner ID is unique across the lifetime of
the program, and is used for for debugging/error detection
(see IDLookupTable.h).

We change GpuDrawItem to store only the actual 16-bit indices,
rather than the 32-bit combination of an index and an 'inner ID'.