Add #define to control enabling of GpuDevice debug code
The GpuDevice validation code that ensures shaders and input layouts
can't be destroyed if there is a pipeline state object referencing them,
and that buffers and pipeline state objects can't be destroyed if there
is a draw item referencing them, will now only be enabled when the
GPUDEVICE_DEBUG_MODE macro is #defined.

This saves memory when the debug code is disabled, as reference counts
for GPU resources (buffers/shaders/etc.) do not need to be stored.

The GpuDevice::RegisterDrawItem()/UnregisterDrawItem() functions will
only be declared when the GPUDEVICE_DEBUG_MODE flag is enabled.

Also, add a GPUDEVICE_UNREGISTER_DRAWITEM(device, item) macro that
calls through to device->UnregisterDrawItem(item) when
GPUDEVICE_DEBUG_MODE is on, and evaluates to a null statement otherwise.