Rename certain GpuDevice API functions
Change the naming convention used for a number of functions in the
GpuDevice class -- in particular, functions for managing GPU resources.

Some examples of the name changes are:
    - CreateShader()               --> ShaderCreate()
    - CreateBuffer()               --> BufferCreate()
    - ShaderIDExists()             --> ShaderExists()
    - GetBufferContents()          --> BufferGetContents()
    - DestroyPipelineStateObject() --> PipelineStateDestroy()
    - MakePerspectiveMatrix()      --> TransformCreatePerspective()

This naming convention change means that the type of resource that the
function operates on is the first element of the function name.

This makes it easier to use autocomplete to find all GpuDevice functions
that operate on a specific resource type (e.g. buffers).