Fix GpuShaderLoad.cpp modifying 'const' data
In GpuShaderLoad::ReadShaderFile(), the 'data' parameter (of type
'const char*') was being cast to a non-const pointer and being modified
(in order to swap the byte-order, if necessary, to handle endianness
differences).

Fix this by copying relevant parts of the data into local stack
variables, which can then be modified.