Fix bug in IDLookupTable::Add()
This fixes an issue in the Add() function which would occur in the case
when the new object was being allocated by reclaiming memory from a
formerly deleted object (in the free list).

In this situation, the new object's ID value wasn't being updated,
causing IDLookupTable::Has() to erroneously return false.

This commit fixes the issue by ensuring that the ID is set correctly.