Fix issue in ShaderAssetFactory::Refresh()
This function was incorrectly passing an argument of '(size_t)size - 1',
rather than just '(size_t)size' to ShaderAsset::Refresh().
(Subtracting 1 was correct in an earlier version of the code that
included a null terminator in the data, but the null terminator had
since been removed and this function had not been updated.)

This commit fixes the issue by passing the correct value of
'(size_t)size' to ShaderAsset::Refresh().