Store model instances' GpuDrawItems in contiguous memory
At present, each model instance consists of a memory allocation that
contains both the ModelInstance data and the corresponding GpuDrawItem.

This commit changes the ModelInstance/ModelRenderQueue classes so that
instead, all the GpuDrawItems for all the ModelInstances are stored in
a single, contiguous allocation.

To achieve this, add a GpuDrawItemPool class, which manages a contiguous
pool of GpuDrawItems and allows structuring GpuDrawItems into
linked lists (implemented in the array).