Implement new & improved model file format
Rather than consisting of a single .mdl file per model, each conceptual
model asset now comprises a .mdl file and a .mdg file.

With this split format, the .mdg file only contains data that is needed
by the CPU during the loading process and is not referenced afterwards.
This is primarily data that is read into memory by the CPU and then
transferred into GPU memory (hence the 'g' in .mdg for GPU). Because of
this, the .mdg file can be read into memory, used during the
model loading process, and then the memory can simply be deallocated
immediately afterwards.

The .mdl file, then, contains data that is used throughout the lifetime
of the application, and so is retained in memory.

The new format also adds support for storing multiple 'submeshes' within
a single model asset. Each submesh can use a different texture, and
all submeshes are drawn when a model is rendered.