Convert to a Z-up coordinate system
This commit changes the world-space coordinate system used in the code
to have the Z-axis pointing upwards, rather than the Y-axis.

The original coordinate system was right-handed and this is still the
case with the new coordinate system.

The .obj model files are assumed to be in a Y-up coordinate system,
and DataCompiler.jar is modified to convert to the Z-up system
when compiling the .obj files.