Add Scene class to handle certain rendering code
This change splits out some of the code previously in the Application
class (related to managing and rendering models). This leaves the
Application class free to focus on handling the user interface.

In addition, this change also decouples the scene rendering code from
the Camera class. The Application class now manages the link between the
two. This is a good idea as the purpose of the Camera class is
conceptually more about the high-level design of the user interface and
user controls for the app, rather than about rendering.