Fix bug and remove workaround in GpuDeviceMetal.mm
A longstanding issue in GpuDeviceMetal.mm was that in the SceneBegin()
member function, -[CAMetalLayer nextDrawable] was occasionally returning
nil (which should not occur). A hack had been developed to work around
this.

The cause of this issue has now been identified -- RenderPassCreate()
in GpuDeviceMetal included an extraneous (i.e.
unnecessary) call to -[CAMetalLayer nextDrawable]. This meant that
an id<CAMetalDrawable> object was never released, which led to
-[CAMetalLayer nextDrawable] later returning nil.

This commit fixes the original issue (the extraneous call to
-[CAMetalLayer nextDrawable]) and removes the workaround.