Improve parameters of GpuDrawItemWriter::Begin()
Change the GpuDrawItemWriter::Begin() function to directly accept
a pointer to the memory in which to write the draw item.

The previous parameters involved passing a function pointer
to an allocator function that GpuDrawItemWriter::Begin() would
use to allocate the memory itself.

This change means that client code of the function should typically
be more straightforward.