Fix synchronization-related issue with 'dynamic' buffers
The GPU_BUFFER_ACCESS_DYNAMIC flag is designed for buffers that are
updated on most frames, and have a fixed maximum number of updates per
frame.

Before this commit, modifying a dynamic buffer would lead to unexpected
results if the buffer already had a draw command in-flight.

This commit resolves the issue without requiring CPU-GPU synchronization
by using, for each logical buffer, a physical buffer containing multiple
copies of the data. Each time an update is made, a different copy is
modified, leaving the GPU to read from the other versions. The number of
copies of the data is equal to the number of frames 'in-flight'
(i.e. 2, for double-buffering) multiplied by the maximum number of
updates per frame.