GpuDeviceMetal: eliminate CPU copy of static buffers
The MTLResourceStorageModeManaged flag was previously being used for
static buffers, which results in a both a CPU and GPU copy of the data.

The CPU copy is redundant as the buffer is never modified and the CPU
never accesses the data.

Eliminate this by first copying the data into a temporary buffer with
'shared' storage mode (MTLResourceStorageModeShared) and then blitting
this into the final buffer, stored only in GPU memory
(MTLResourceStorageModePrivate).