Mac: Don't copy Assets folder into the app bundle
The current version of the code always copies the Assets directory
into the app bundle's Resources folder.

This commit changes this so that instead:

- The Assets folder is >not< copied into the bundle, and instead:

- A symbolic link to the Assets folder is created in the same directory
  as the app bundle.

- The app's working directory is set to the bundle's parent directory,
  rather than the bundle Resources folder as before.

Using a symbolic link instead of a copy improves build performance and
removes the redundancy of the data.

In addition, placing the symbolic link alongside the app bundle (in its
parent directory), rather in the bundle Resources folder, will improve
the consistency with other operating systems/platforms.