Add a 'null' GpuDevice implementation
This commit adds a new file, GpuDeviceNull.cpp, that implements all the
functions in GpuDevice.h, but does not display anything and does not use
a graphics API.

This is intended to be useful for debugging or as a starting point when
developing actual renderers for different graphics APIs.

The commit updates premake5.lua to allow switching between the GpuDevice
implementations.